+++ lib/eda/api/thread.ex
@@ -23,7 +23,7 @@ defmodule EDA.API.Thread do
           {:ok, map()} | {:error, term()}
   def start_from_message(channel_id, message_id, opts) do
     body = Map.new(opts)
-    check_options(body, @from_message_keys, "EDA.API.Thread.start_from_message/3")
+    check_options!(body, @from_message_keys, "EDA.API.Thread.start_from_message/3")
     post("/channels/#{channel_id}/messages/#{message_id}/threads", body)
   end
 
@@ -31,7 +31,7 @@ defmodule EDA.API.Thread do
   @spec start(String.t() | integer(), map() | keyword()) :: {:ok, map()} | {:error, term()}
   def start(channel_id, opts) do
     body = Map.new(opts)
-    check_options(body, @thread_keys, "EDA.API.Thread.start/2")
+    check_options!(body, @thread_keys, "EDA.API.Thread.start/2")
     post("/channels/#{channel_id}/threads", body)
   end
 
@@ -80,8 +80,8 @@ defmodule EDA.API.Thread do
   @spec create_post(String.t() | integer(), keyword(), keyword()) ::
           {:ok, map()} | {:error, term()}
   def create_post(channel_id, opts, message_opts \\ []) do
-    check_options(opts, @forum_keys, "EDA.API.Thread.create_post/3")
-    check_options(message_opts, @post_message_keys, "EDA.API.Thread.create_post/3 (message)")
+    check_options!(opts, @forum_keys, "EDA.API.Thread.create_post/3")
+    check_options!(message_opts, @post_message_keys, "EDA.API.Thread.create_post/3 (message)")
     path = "/channels/#{channel_id}/threads"
 
     case build_message_payload(message_opts) do